+++ src/store/getters.js
@@ -1,0 +1,8 @@
+const storedMemoList = (state) => {
+  return state.memoList;
+};
+const storedCurrentData = (state) => {
+  return state.currentData;
+};
+
+export { storedCurrentData, storedMemoList };
